+++ common/src/main/java/tocraft/walkers/api/blacklist/EntityBlacklist.java
@@ -1,18 +1,28 @@
 package tocraft.walkers.api.blacklist;
 
+import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.EntityType;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class EntityBlacklist {
-    private static final List<EntityType<?>> blacklist = new ArrayList<>();
+    private static final List<EntityType<?>> typeBlacklist = new ArrayList<>();
+    private static final List<TagKey<EntityType<?>>> tagBlacklist = new ArrayList<>();
 
     public static boolean isBlacklisted(EntityType<?> entityType) {
-        return blacklist.contains(entityType);
+        if (typeBlacklist.contains(entityType)) return true;
+        for (TagKey<EntityType<?>> entityTypeTagKey : tagBlacklist) {
+            if (entityType.is(entityTypeTagKey)) return true;
+        }
+        return false;
     }
 
     public static void registerByType(EntityType<?> entityType) {
-        blacklist.add(entityType);
+        typeBlacklist.add(entityType);
+    }
+
+    public static void registerByTag(TagKey<EntityType<?>> entityTypeTag) {
+        tagBlacklist.add(entityTypeTag);
     }
 }