+++ common/src/main/java/tocraft/walkers/api/data/blacklist/EntityBlacklistDataManager.java
@@ -1,19 +1,22 @@
 package tocraft.walkers.api.data.blacklist;
 
 import com.google.gson.*;
+import com.mojang.datafixers.util.Pair;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.JsonOps;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 import net.minecraft.Util;
 import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.packs.resources.ResourceManager;
 import net.minecraft.server.packs.resources.SimpleJsonResourceReloadListener;
+import net.minecraft.tags.TagKey;
 import net.minecraft.util.profiling.ProfilerFiller;
-import net.minecraft.world.entity.EntityType;
 import tocraft.walkers.Walkers;
 import tocraft.walkers.api.blacklist.EntityBlacklist;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
@@ -24,24 +27,28 @@ public class EntityBlacklistDataManager extends SimpleJsonResourceReloadListener
         super(GSON, Walkers.MODID);
     }
 
-    @SuppressWarnings("unchecked")
     @Override
     protected void apply(Map<ResourceLocation, JsonElement> map, ResourceManager resourceManager, ProfilerFiller profiler) {
         for (Map.Entry<ResourceLocation, JsonElement> mapEntry : map.entrySet()) {
             if (mapEntry.getKey().getPath().equals("blacklist")) {
-                List<? extends EntityType<?>> someBlacklist = blacklistFromJson(mapEntry.getValue().getAsJsonObject());
-
-                for (EntityType<?> entityType : someBlacklist) {
-                    EntityBlacklist.registerByType(entityType);
+                Pair<List<ResourceLocation>, List<ResourceLocation>> someBlacklist = blacklistFromJson(mapEntry.getValue().getAsJsonObject());
+                for (ResourceLocation resourceLocation : someBlacklist.getFirst()) {
+                    if (BuiltInRegistries.ENTITY_TYPE.containsKey(resourceLocation)) {
+                        EntityBlacklist.registerByType(BuiltInRegistries.ENTITY_TYPE.get(resourceLocation));
+                    }
+                }
+                for (ResourceLocation resourceLocation : someBlacklist.getSecond()) {
+                    EntityBlacklist.registerByTag(TagKey.create(Registries.ENTITY_TYPE, resourceLocation));
                 }
             }
         }
     }
 
-    protected static List<? extends EntityType<?>> blacklistFromJson(JsonObject json) {
-        Codec<List<? extends EntityType<?>>> codec = RecordCodecBuilder.create((instance) -> instance.group(
-                Codec.list(ResourceLocation.CODEC).fieldOf("entity_types").forGetter(o -> null)
-        ).apply(instance, instance.stable((entityTypes) -> entityTypes.stream().filter(type -> BuiltInRegistries.ENTITY_TYPE.containsKey(type)).map(type -> BuiltInRegistries.ENTITY_TYPE.get(type)).toList())));
+    protected static Pair<List<ResourceLocation>, List<ResourceLocation>> blacklistFromJson(JsonObject json) {
+        Codec<Pair<List<ResourceLocation>, List<ResourceLocation>>> codec = RecordCodecBuilder.create((instance) -> instance.group(
+                Codec.list(ResourceLocation.CODEC).optionalFieldOf("entity_types", new ArrayList<>()).forGetter(Pair::getFirst),
+                Codec.list(ResourceLocation.CODEC).optionalFieldOf("entity_tags", new ArrayList<>()).forGetter(Pair::getSecond)
+        ).apply(instance, instance.stable(Pair::new)));
         return Util.getOrThrow(codec.parse(JsonOps.INSTANCE, json), JsonParseException::new);
     }
 }