+++ common/src/main/java/tocraft/walkers/skills/SkillRegistry.java
@@ -3,6 +3,7 @@ package tocraft.walkers.skills;
 import com.mojang.serialization.Codec;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.FluidTags;
+import net.minecraft.tags.TagKey;
 import net.minecraft.world.damagesource.DamageTypes;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.effect.MobEffects;
@@ -111,6 +112,7 @@ public class SkillRegistry {
         // reinforcements
         registerByClass(Wolf.class, new ReinforcementsSkill<>());
         registerByClass(Bee.class, new ReinforcementsSkill<>());
+        //registerByTag(EntityTypeTags.RAIDERS, new ReinforcementsSkill<>(32, new ArrayList<>(), List.of(EntityTypeTags.RAIDERS)));
         // instant die on lightning
         registerByClass(Turtle.class, new InstantDieOnDamageTypeSkill<>(DamageTypes.LIGHTNING_BOLT));
         // cats hunt rabbits
@@ -153,9 +155,12 @@ public class SkillRegistry {
         return skillList;
     }
 
-
     public static <A extends LivingEntity> void registerByType(EntityType<A> type, ShapeSkill<A> skill) {
         registerByPredicate(livingEntity -> type.equals(livingEntity.getType()), skill);
+    }
+
+    public static <A extends LivingEntity> void registerByTag(TagKey<EntityType<?>> tag, ShapeSkill<A> skill) {
+        registerByPredicate(livingEntity -> livingEntity.getType().is(tag), skill);
     }
 
     public static <A extends LivingEntity> void registerByClass(Class<A> entityClass, ShapeSkill<A> skill) {