+++ common/src/main/java/tocraft/walkers/skills/impl/PreySkill.java
@@ -3,7 +3,9 @@ package tocraft.walkers.skills.impl;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
 import tocraft.walkers.Walkers;
@@ -19,13 +21,15 @@ public class PreySkill<E extends LivingEntity> extends ShapeSkill<E> {
     public static final Codec<PreySkill<?>> CODEC = RecordCodecBuilder.create((instance) -> instance.group(
             Codec.list(ResourceLocation.CODEC).fieldOf("hunter").forGetter(o -> new ArrayList<>())
     ).apply(instance, instance.stable(hunterLocations -> {
-                List<EntityType<?>> hunterTypes = new ArrayList<>();
+                List<Predicate<LivingEntity>> hunter = new ArrayList<>();
                 for (ResourceLocation resourceLocation : hunterLocations) {
                     if (BuiltInRegistries.ENTITY_TYPE.containsKey(resourceLocation)) {
-                        hunterTypes.add(BuiltInRegistries.ENTITY_TYPE.get(resourceLocation));
+                        hunter.add(entity -> entity.getType().equals(BuiltInRegistries.ENTITY_TYPE.get(resourceLocation)));
+                    } else {
+                        hunter.add(entity -> entity.getType().is(TagKey.create(Registries.ENTITY_TYPE, resourceLocation)));
                     }
                 }
-                return ofHunterType(hunterTypes.toArray(EntityType[]::new));
+                return new PreySkill<>(hunter);
             }
     )));
 
@@ -33,6 +37,11 @@ public class PreySkill<E extends LivingEntity> extends ShapeSkill<E> {
 
     public static PreySkill<?> ofHunterType(EntityType<?>... hunter) {
         return new PreySkill<>(Stream.of(hunter).map(entry -> (Predicate<LivingEntity>) entity -> entity.getType().equals(entry)).toList());
+    }
+
+    @SafeVarargs
+    public static PreySkill<?> ofHunterTag(TagKey<EntityType<?>>... hunter) {
+        return new PreySkill<>(Stream.of(hunter).map(entry -> (Predicate<LivingEntity>) entity -> entity.getType().is(entry)).toList());
     }
 
     @SafeVarargs